cases/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-jalali-date==1.0.2
jdatetime==4.1.1
mysqlclient==2.2.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
